Split StringField into single/multi value components
StringField handled both value modes in one component via a ternary,
with the multi-only tag-list state and handlers sitting inline next
to the single-value path, and useField called against a
string | string[] union that forced a cast on the single-value input.

Splitting into SingleStringField and MultiStringField, dispatched
from a thin StringField, isolates each value shape to its own
useField call so format and parse are typed against the correct
shape without a cast, and the tag-list logic no longer lives next to
code that never touches it. StringFieldProps is now a discriminated
union on `multi` to support this.

Also inlines StringTag's getTagRootStyle, which only had two call
sites in the same file and didn't need to be a separate function.
