+++ javascript/packages/core/components/form/fields/string/types.ts
@@ -1,5 +1,6 @@
 import type { BaseFieldProps } from '../types';
 
-export type StringFieldProps =
-  | (BaseFieldProps<string> & { multi?: false })
-  | (BaseFieldProps<string[]> & { multi: true });
+export type SingleStringFieldProps = BaseFieldProps<string> & { multi?: false };
+export type MultiStringFieldProps = BaseFieldProps<string[]> & { multi: true };
+
+export type StringFieldProps = SingleStringFieldProps | MultiStringFieldProps;
